Classifier created in Keras for rust detection
Classifier created in Keras to detect and classify images and whether they create rust in them. Trained on a sample of data from bridge inspection records and can be used to detect whether images contain rust with the intent of also loading it onto a mobile device to use on device labelling of images in the field (hence the use of MobileNets as well as Xception).